api/requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12          # multipart/form-data file uploads
pydantic==2.9.2
pydantic-settings==2.6.1

# ── Image Processing ──────────────────────────────────────────────────────────
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
rembg==2.0.59                     # background removal (U2Net / RMBG)
onnxruntime-gpu==1.19.2           # GPU-accelerated onnx (used by rembg)

# ── Deep Learning Runtime ──────────────────────────────────────────────────────
torch==2.4.1                      # install separately via: pip install torch --index-url https://download.pytorch.org/whl/cu121
torchvision==0.19.1
accelerate==0.34.2
diffusers==0.30.3                 # for OOTDiffusion / CatVTON
transformers==4.44.2
huggingface_hub==0.25.1

# ── 3D / Mesh ──────────────────────────────────────────────────────────────────
trimesh==4.4.9                    # mesh I/O, cleaning, export (.obj / .glb)
pyglet==1.5.29                    # trimesh optional viewer (headless OK without it)
scipy==1.14.1                     # trimesh dependency

# ── Utilities ──────────────────────────────────────────────────────────────────
aiofiles==24.1.0                  # async file I/O
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2                     # async HTTP (inter-service calls if needed)
